Test replacing lines with printf and signals
The printf thing could be used in the completion part of minishell to
rewrite the line when switching between possibilities with tabulation.
Signals will have to be used basically everywhere. The input part of
minishell should be its own process so that it can be Ctrl-C'd and
Ctrl-D (for when this will be handled).