Flush msd with a static variable for Ctrl-C
This also adds the Makefile. Anyway, signal handling in ms_input doesn't
work. This is because it waits until read is finished. Maybe ms_input
should indeed be its own process, but the input buffer will have to be
transferred from ms_input to minishell, maybe via a bonus function
(pipe?). But it compiles, so that's that I guess.