Collect every malloc and handle signals inside it
I am doing something very hacky in ft_secmalloc. I am casting an
int(*)(int) into a void(*)(int) to feed to the signal function. This
allows me to get the value of the static being stored inside the signal
handling function I give to signal. It looks like it works and the
compiler doesn't protest, but this is dark magic. I am doing this to
avoid a SIGINT during a malloc, which could end up in leaks (yeah, I am
a bit paranoid about it since my ft_ls was given a 0 for this very
reason).