Read input as multibyte characters
This allows to detect special keys like arrows or home/end and
page_up/page_down. But input is kind of buggy right now (more like
incomplete really). This will have to be redone and organised well.
Because right now, this is ugly. That's why I added a TODO, this is a
great tool to assist my barely functionning brain.