Keep cursor movement coherent
Now there is a boolean (bol for "begining of line") in ms_input
that indicates if we are at the begining of the line or not. This
prevents the cursor from moving outside of the boundries of input
when placing it at the begining of the line. Now the move_left and
move_right functions (which correspond to ARROW_LEFT and ARROW_RIGHT
input) are working. The other functions (especially insert_char and
back_delete) will have to be updated to take this into account.