Reset terminal attributes with ft_atexit
This is a update to the last version of the libft. This includes
the ft_exit, ft_atexit and ft_exitmsg functions and a generic
version of the heap_collector. This will allow the libft to catch
open files and directories to close them when the program exits,
for example. This version of the libft works well, but it still
has not been tested for leaks.

Now the terminal attributes are set in the reset_input_mode funtion.
The funtction is called back at the end of the program with ft_atexit.
This restores the terminal cleanly at exit.