Cut the input_buffer cleanly
This fixes the problems introduced in the last commit. Now the
parser is able to cut the string and handle ';'. The command
strings can be empty if there is multiple ';' in a row, or if there
are only spaces between them. The recursive function also works.
Further testing needs to be done though, to make sure that there is
no leak in this part of the program.