Delete parsing and execution
This is too early to do them. I must have something totally clean
for the input part first. Also the prototype of parsing was going
in the wrong direction (the problem is way more complex than what
my dumb split solution could handle). The prototype of exectution
was only printing the result of parsing, so not much was lost.