Remove quotes in parsing
This is the final transformation to be done on input before the
execution phase. The removing of quotes is based on get_qmode
function. It duplicates the current string if it finds quotes to
be removed, and it just does not copy the quotes. This looks like
it works, but it needs further testing.