Execute a program with a temporary environment
This is what the env builtin does. If it is call without any argument
it will only print the environment, if it is called with arguments
of the form VAR=value and a command after it, it will execute with
the environment variables given (replacing already existing ones if
need be). With the option -i, it can give an empty environment to
the command. This also works if variables are added after the option.