Fix join_lines and qmode
When removing a '\' fallowed by a '\n' in join_lines, a character
the next character was ignored, so if it was a quote it wasnt taken
into account. This gives a wrong input to the lexer.

Also, the quote modes were defined twice. Once in quotes.h and in
ms_input.h which leaded to a wrong value for qmode (this also
resulted in ignored '\').