Complete input when pressing tab
This works very well with the executables. With files it only works
if there is a '/' (at least to the best of my knowledge), and when
completing an env variable it apparently just crashes. I am still
commiting now because I did a lot of work at once, and I gotta say
I am surprised it even half works.