Fix word count around ';'
The dllst and regular word count have been fixed on this. Before,
if a word was next to a ';' without blank characters in between,
it did not count the new word.

Also, now when counting a word in the regular version, it does
not count a trailing '\' and '\n' as a word. This change has
not been made in the dllst version because it should never run
into a '\n' in the middle of an input string.