Count commands from minishell startup
This will be used to only add new commands to the history, not what
has been loaded at startup. Although maybe this wont be useful in
the end. The flush_history function will probably be split in two
because it is hard to do all needed operations in one 25 lines
function.