Return 1 by default on commands
I am sorry to say that, but "the default is failure". This is just
because it is easier to set the cmd_exit to 1 before executing
than after (because of the 25 lines limit). Also, the erasing of
lines will have to be redone in the input, because it does not delete
the lines (if there are '\n' in the input).