Use back_delete in erase_line and handle new lines
The handling of back deletion is better now, but it is still problematic
when the line is exactly as long as the width of the terminal and the
cursor is "outside" the terminal (it is half under the scroll bar). This
won't be extra easy to detect but it can be. This should be fixed.