Protect the opening and reading of history file
This avoid opening a symlink as the history. It also prevents the
load_history function from attempting to allocate a gigantic buffer
if the history file has been corrupted. CMDMAX is used to limit the
size of commands registered by the history, so if one is superior
to it, then it stops reading the .minihistory file.